requirements/base.txt
# Primary requirements
Django== 3.2.17
drf-rw-serializers==1.0.5
drf-model-serializer==0.0.2
djangorestframework==3.14.0
requests==2.28.2
python-decouple==3.1

# Django Model Field that uses Hashids to obscure the value
# https://github.com/nshafer/django-hashid-field
django-hashid-field==3.1.3

django-postgres-extra==2.0.6

# psycopg2-binary
# https://pypi.org/project/psycopg2-binary
psycopg2-binary==2.9.3

django-filter==22.1

